build: update Vercel build output files and client bundles
